requirements.txt
Flask==2.0.2
Pillow==8.1.2
Pillow==9.5.0
Werkzeug==2.0.2
